requirements.txt
openai
pynput
sounddevice
numpy
openai-whisper
playwright
pyautogui
python-dotenv
